remove storageClassClaim and add new cr
from 4.13 we are moving from storageClassClaim
to storageClassRequest CR. Also, we don't need
consumer mode code in ocs-operator now, new operator
will have  consumer.

(cherry picked from commit 55e5c082d52ab5ec27c5ff151ab2392c1023bccb)
