+++ config/crd/patches/cainjection_in_storageclassrequests.yaml
@@ -5,4 +5,4 @@ kind: CustomResourceDefinition
 metadata:
   annotations:
     cert-manager.io/inject-ca-from: $(CERTIFICATE_NAMESPACE)/$(CERTIFICATE_NAME)
-  name: storageclassclaims.ocs.openshift.io
+  name: storageclassrequests.ocs.openshift.io
