+++ config/crd/patches/webhook_in_storageclassrequests.yaml
@@ -3,7 +3,7 @@
 apiVersion: apiextensions.k8s.io/v1beta1
 kind: CustomResourceDefinition
 metadata:
-  name: storageclassclaims.ocs.openshift.io
+  name: storageclassrequests.ocs.openshift.io
 spec:
   conversion:
     strategy: Webhook
