--- config/rbac/storageclassclaim_editor_role.yaml
@@ -1,24 +1,0 @@
-# permissions for end users to edit storageclassclaims.
-apiVersion: rbac.authorization.k8s.io/v1
-kind: ClusterRole
-metadata:
-  name: storageclassclaim-editor-role
-rules:
-- apiGroups:
-  - ocs.openshift.io
-  resources:
-  - storageclassclaims
-  verbs:
-  - create
-  - delete
-  - get
-  - list
-  - patch
-  - update
-  - watch
-- apiGroups:
-  - ocs.openshift.io
-  resources:
-  - storageclassclaims/status
-  verbs:
-  - get
