--- config/rbac/storageclassclaim_viewer_role.yaml
@@ -1,20 +1,0 @@
-# permissions for end users to view storageclassclaims.
-apiVersion: rbac.authorization.k8s.io/v1
-kind: ClusterRole
-metadata:
-  name: storageclassclaim-viewer-role
-rules:
-- apiGroups:
-  - ocs.openshift.io
-  resources:
-  - storageclassclaims
-  verbs:
-  - get
-  - list
-  - watch
-- apiGroups:
-  - ocs.openshift.io
-  resources:
-  - storageclassclaims/status
-  verbs:
-  - get
