+++ config/samples/ocs_v1alpha1_storageclassrequest.yaml
@@ -1,7 +1,7 @@
 apiVersion: ocs.openshift.io/v1alpha1
-kind: StorageClassClaim
+kind: StorageClassRequest
 metadata:
-  name: storageclassclaim-sample
+  name: storageclassrequest-sample
 spec:
   # Add fields here
   foo: bar
