+++ services/provider/server/server.go
@@ -33,16 +33,16 @@ import (
 	"k8s.io/apimachinery/pkg/api/resource"
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/types"
-	"k8s.io/klog"
+	klog "k8s.io/klog/v2"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/client/config"
 )
 
 const (
-	TicketAnnotation           = "ocs.openshift.io/provider-onboarding-ticket"
-	ProviderCertsMountPoint    = "/mnt/cert"
-	onboardingTicketKeySecret  = "onboarding-ticket-key"
-	storageClassClaimNameLabel = "ocs.openshift.io/storageclassclaim-name"
+	TicketAnnotation             = "ocs.openshift.io/provider-onboarding-ticket"
+	ProviderCertsMountPoint      = "/mnt/cert"
+	onboardingTicketKeySecret    = "onboarding-ticket-key"
+	storageClassRequestNameLabel = "ocs.openshift.io/storageclassrequest-name"
 )
 
 const (
@@ -52,10 +52,10 @@ const (
 
 type OCSProviderServer struct {
 	pb.UnimplementedOCSProviderServer
-	client                   client.Client
-	consumerManager          *ocsConsumerManager
-	storageClassClaimManager *storageClassClaimManager
-	namespace                string
+	client                     client.Client
+	consumerManager            *ocsConsumerManager
+	storageClassRequestManager *storageClassRequestManager
+	namespace                  string
 }
 
 type onboardingTicket struct {
@@ -74,16 +74,16 @@ func NewOCSProviderServer(ctx context.Context, namespace string) (*OCSProviderSe
 		return nil, fmt.Errorf("failed to create new OCSConumer instance. %v", err)
 	}
 
-	storageClassClaimManager, err := newStorageClassClaimManager(ctx, client, namespace)
+	storageClassRequestManager, err := newStorageClassRequestManager(ctx, client, namespace)
 	if err != nil {
 		return nil, fmt.Errorf("failed to create new StorageClassClaim instance. %v", err)
 	}
 
 	return &OCSProviderServer{
-		client:                   client,
-		consumerManager:          consumerManager,
-		storageClassClaimManager: storageClassClaimManager,
-		namespace:                namespace,
+		client:                     client,
+		consumerManager:            consumerManager,
+		storageClassRequestManager: storageClassRequestManager,
+		namespace:                  namespace,
 	}, nil
 }
 
@@ -398,7 +398,7 @@ func (s *OCSProviderServer) getCephClientInformation(ctx context.Context, name s
 	if len(cephClient.Annotations) == 0 {
 		return "", "", fmt.Errorf("rook ceph client %s annotation is empty", name)
 	}
-	if cephClient.Annotations[controllers.StorageClaimAnnotation] == "" || cephClient.Annotations[controllers.StorageCephUserTypeAnnotation] == "" {
+	if cephClient.Annotations[controllers.StorageRequestAnnotation] == "" || cephClient.Annotations[controllers.StorageCephUserTypeAnnotation] == "" {
 		klog.Warningf("rook ceph client %s has missing storage annotations", name)
 	}
 
@@ -506,9 +506,9 @@ func (s *OCSProviderServer) FulfillStorageClassClaim(ctx context.Context, req *p
 		return nil, status.Errorf(codes.InvalidArgument, "encountered an unknown storage type, %s", storageType)
 	}
 
-	err = s.storageClassClaimManager.Create(ctx, consumerObj, req.StorageClassClaimName, storageType, req.EncryptionMethod, req.StorageProfile)
+	err = s.storageClassRequestManager.Create(ctx, consumerObj, req.StorageClassClaimName, storageType, req.EncryptionMethod, req.StorageProfile)
 	if err != nil {
-		errMsg := fmt.Sprintf("failed to fulfill storage class claim for %q. %v", req.StorageConsumerUUID, err)
+		errMsg := fmt.Sprintf("failed to fulfill storage class request for %q. %v", req.StorageConsumerUUID, err)
 		klog.Error(errMsg)
 		if kerrors.IsAlreadyExists(err) {
 			return nil, status.Error(codes.AlreadyExists, errMsg)
@@ -522,9 +522,9 @@ func (s *OCSProviderServer) FulfillStorageClassClaim(ctx context.Context, req *p
 // RevokeStorageClassClaim RPC call to delete the StorageclassClaim CR on
 // provider cluster.
 func (s *OCSProviderServer) RevokeStorageClassClaim(ctx context.Context, req *pb.RevokeStorageClassClaimRequest) (*pb.RevokeStorageClassClaimResponse, error) {
-	err := s.storageClassClaimManager.Delete(ctx, req.StorageConsumerUUID, req.StorageClassClaimName)
+	err := s.storageClassRequestManager.Delete(ctx, req.StorageConsumerUUID, req.StorageClassClaimName)
 	if err != nil {
-		errMsg := fmt.Sprintf("failed to revoke storage class claim %q for %q. %v", req.StorageClassClaimName, req.StorageConsumerUUID, err)
+		errMsg := fmt.Sprintf("failed to revoke storage class request %q for %q. %v", req.StorageClassClaimName, req.StorageConsumerUUID, err)
 		klog.Error(errMsg)
 		return nil, status.Errorf(codes.Internal, errMsg)
 	}
@@ -534,9 +534,9 @@ func (s *OCSProviderServer) RevokeStorageClassClaim(ctx context.Context, req *pb
 
 // GetStorageClassClaim RPC call to get the ceph resources for the StorageclassClaim.
 func (s *OCSProviderServer) GetStorageClassClaimConfig(ctx context.Context, req *pb.StorageClassClaimConfigRequest) (*pb.StorageClassClaimConfigResponse, error) {
-	storageClassClaim, err := s.storageClassClaimManager.Get(ctx, req.StorageConsumerUUID, req.StorageClassClaimName)
+	storageClassClaim, err := s.storageClassRequestManager.Get(ctx, req.StorageConsumerUUID, req.StorageClassClaimName)
 	if err != nil {
-		errMsg := fmt.Sprintf("failed to get storage class claim config %q for %q. %v", req.StorageClassClaimName, req.StorageConsumerUUID, err)
+		errMsg := fmt.Sprintf("failed to get storage class request config %q for %q. %v", req.StorageClassClaimName, req.StorageConsumerUUID, err)
 		if kerrors.IsNotFound(err) {
 			return nil, status.Error(codes.NotFound, errMsg)
 		}
@@ -544,18 +544,18 @@ func (s *OCSProviderServer) GetStorageClassClaimConfig(ctx context.Context, req 
 	}
 
 	// Verify Status.Phase
-	msg := fmt.Sprintf("storage class claim %q for %q is in %q phase", req.StorageClassClaimName, req.StorageConsumerUUID, storageClassClaim.Status.Phase)
+	msg := fmt.Sprintf("storage class request %q for %q is in %q phase", req.StorageClassClaimName, req.StorageConsumerUUID, storageClassClaim.Status.Phase)
 	klog.Info(msg)
-	if storageClassClaim.Status.Phase != ocsv1alpha1.StorageClassClaimReady {
+	if storageClassClaim.Status.Phase != ocsv1alpha1.StorageClassRequestReady {
 		switch storageClassClaim.Status.Phase {
-		case ocsv1alpha1.StorageClassClaimFailed:
+		case ocsv1alpha1.StorageClassRequestFailed:
 			return nil, status.Error(codes.Internal, msg)
-		case ocsv1alpha1.StorageClassClaimInitializing:
+		case ocsv1alpha1.StorageClassRequestInitializing:
 			return nil, status.Error(codes.Unavailable, msg)
-		case ocsv1alpha1.StorageClassClaimCreating:
+		case ocsv1alpha1.StorageClassRequestCreating:
 			return nil, status.Error(codes.Unavailable, msg)
 		case "":
-			return nil, status.Errorf(codes.Unavailable, "status is not set for storage class claim %q for %q", req.StorageClassClaimName, req.StorageConsumerUUID)
+			return nil, status.Errorf(codes.Unavailable, "status is not set for storage class request %q for %q", req.StorageClassClaimName, req.StorageConsumerUUID)
 		default:
 			return nil, status.Error(codes.Internal, msg)
 		}
@@ -672,7 +672,7 @@ func (s *OCSProviderServer) GetStorageClassClaimConfig(ctx context.Context, req 
 				})})
 		}
 	}
-	klog.Infof("successfully returned the storage class claim %q for %q", req.StorageClassClaimName, req.StorageConsumerUUID)
+	klog.Infof("successfully returned the storage class request %q for %q", req.StorageClassClaimName, req.StorageConsumerUUID)
 	return &pb.StorageClassClaimConfigResponse{ExternalResource: extR}, nil
 
 }
